+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - `fullscreen` rule to hide notifications when a fullscreen window is active
+- When new notifications arrive, but display is full, important notifications don't
+  have to wait for a timeout in a displayed notification (#541)
 
 ## 1.3.2 - 2018-05-06